fitur autentikasi (login, register, dan logout), serta mengecek status user yang terautentikasi

+++ app/Http/Controllers/DeviceController.php
@@ -5,6 +5,7 @@ namespace App\Http\Controllers;
 use App\Http\Requests\DeviceRequest;
 use App\Models\Device;
 use Illuminate\Http\Request;
+use Illuminate\Support\Facades\Auth;
 
 class DeviceController extends Controller
 {
@@ -21,6 +22,10 @@ class DeviceController extends Controller
 
     public function create()
     {
+        if (!Auth::check()) {
+            return back();
+        }
+
         return view('devices.create');
     }
 
@@ -47,6 +52,10 @@ class DeviceController extends Controller
 
     public function edit($id)
     {
+        if (!Auth::check()) {
+            return back();
+        }
+
         $device = Device::find($id);
 
         return view('devices.edit', [
@@ -76,6 +85,10 @@ class DeviceController extends Controller
     }
 
     public function delete($id) {
+        if (!Auth::check()) {
+            return back();
+        }
+
         $device = Device::where('id', $id);
         $device->delete();
